# Conflicts:
#	token-common/src/main/java/com/token/constant/MessageConstant.java
#	token-server/src/main/java/com/token/controller/admin/SetmealController.java
#	token-server/src/main/java/com/token/mapper/GoodsMapper.java
#	token-server/src/main/java/com/token/mapper/SetmealMapper.java
#	token-server/src/main/java/com/token/service/GoodsService.java
#	token-server/src/main/java/com/token/service/SetmealService.java
#	token-server/src/main/java/com/token/service/impl/EmployeeServiceImpl.java
#	token-server/src/main/java/com/token/service/impl/GoodsServiceImpl.java
#	token-server/src/main/resources/mapper/EmployeeMapper.xml
#	token-server/src/main/resources/mapper/SetmealMapper.xml

+++ token-server/src/main/java/com/token/service/impl/GoodsServiceImpl.java
@@ -94,18 +94,20 @@ public class GoodsServiceImpl implements GoodsService {
     }
 
     /**
-     * 删除商品
+     * 根据id批量删除
      * @param ids
      */
     public void delete(Long[] ids) {
+        List<Long> stauts = goodsMapper.getGoodsByStatusId(ids);
+        if (stauts.size() > 0) {
+            throw new AccountIsDisableException(MessageConstant.GOODS_STATUS_IS_ENABLE);
+        }
         goodsMapper.delete(ids);
     }
 
     /**
      * 根据id查询商品信息
-     *
      * @param id
-     * @return
      */
     public Goods getGoodsInfo(Long id) {
         return goodsMapper.getGoodsById(id);
@@ -121,7 +123,8 @@ public class GoodsServiceImpl implements GoodsService {
     }
 
     /**
-     *修改商品状态
+     * 修改商品状态
+     *
      * @param id
      * @param status
      */
